+++ src/components/CityComponents/HostSensors/HostCard.js
