+++ src/pages/air/City.js
@@ -6,10 +6,10 @@ import KenyaMap from "../../components/Maps/KenyaMap";
 
 import Neighbourhood from "../../components/CityComponents/Neighbourhood";
 import StatisticsReport from "../../components/CityComponents/StatisticsReport";
-import CityHeader from "../../components/CityComponents/CityHeader";
+import CityHeader from "../../components/CityComponents/Header/CityHeader";
 import CallToAction from "../../components/CityComponents/CallToAction";
-import SensorsDataReport from "../../components/CityComponents/SensorsDataReport";
-import HostSensor from "../../components/CityComponents/HostSensor";
+import SensorsDataReport from "../../components/CityComponents/MeasureStats/SensorsDataReport";
+import HostSensor from "../../components/CityComponents/HostSensors/HostSensor";
 
 //import Api from "../../components/Api";
 //import Typography from "@material-ui/core/Typography";
